Add .gitignore file for Python and JetBrains IDEs
This update introduces a .gitignore file to untrack unnecessary or redundant files in the project. The file is configured to ignore common Python files (e.g. .pyc, __pycache__), PyCharm editor files, Django and Flask files, and more that should not be part of source control. This helps to keep the repository clean and focused on meaningful project files.